Fix monitor mode not working with IPv6, sockets or lua scripts
Fixes #1189
Fixes #1037
